requirements.txt
django==2.1.3
djangorestframework==3.9.0
django-import-export==1.1.0
django-filter==2.0.0
